apply uniform API

+++ source/runtime/ppp_engine/private/image.cpp
@@ -61,11 +61,11 @@ namespace ppp
                 return false;
             }
 
-            u64 vertex_count() const override
+            u32 vertex_count() const override
             {
                 return m_geometry->vertex_count();
             }
-            u64 index_count() const override
+            u32 index_count() const override
             {
                 return m_geometry->index_count();
             }
